상품조회

+++ src/main/java/org/test/teamproject_back/dto/request/ReqAddProductDto.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class ReqAddProductDto {
     private String title;
-    private long price;
-    private long stock;
+    private int price;
+    private int stock;
     private int categoryId;
     private String origin;
     private String img;
